ai应用部分代码/feima-ai-service/requirements.txt
# Web
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sse-starlette>=1.8.0

# LangChain
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# HTTP
httpx>=0.27.0
requests>=2.31.0

# OpenAI 兼容客户端（火山引擎豆包等）
openai>=1.0.0

# Data
pydantic>=1.10.0
python-dotenv>=1.0.0

# Database (MySQL, optional when using Java API)
SQLAlchemy>=2.0.0
aiomysql>=0.2.0
